+++ cmd/foreverdubbed/main.go
@@ -23,7 +23,7 @@ import (
 	"foreverdubbed/internal/speech"
 )
 
-const version = "0.6.5"
+const version = "0.6.6"
 
 func main() {
 	prepareConsole()